requirement.txt
fastmcp==2.14.3
langchain==1.2.6
langchain-google-genai==4.2.0
langchain-community==0.4.1
python-dotenv==1.2.1
requests==2.32.5